+++ modules/planning/math/smoothing_spline/spline_1d_generator.cc
@@ -29,7 +29,7 @@ namespace apollo {
 namespace planning {
 
 Spline1dGenerator::Spline1dGenerator(const std::vector<double>& x_knots,
-                                     const std::size_t spline_order)
+                                     const std::uint32_t spline_order)
     : spline_(x_knots, spline_order),
       spline_constraint_(x_knots, spline_order),
       spline_kernel_(x_knots, spline_order) {}
@@ -73,7 +73,7 @@ bool Spline1dGenerator::solve() {
     return false;
   }
 
-  const std::size_t spline_order = spline_.spline_order();
+  const std::uint32_t spline_order = spline_.spline_order();
   const Eigen::MatrixXd solved_params = qp_solver_->params();
   return spline_.set_spline_segs(solved_params, spline_order);
 }
